feat: author_id 구현, users 목록 불러오기

+++ app/services/post_service.py
@@ -18,7 +18,7 @@ class PostService:
         ).all()
         return posts
         
-    def create_post(self, db:Session, post: PostReq):
+    def create_post(self, db:Session, post: PostReq, userDict: dict):
         postModel = Post()
         postModel.title = post.title
         postModel.body = post.body
@@ -27,21 +27,25 @@ class PostService:
         postModel.created_at = int(time.time())
         postModel.location = post.location
         postModel.updated_at = postModel.created_at
+
+        user_id = userDict.get('id', 0)
+        postModel.author_id = user_id
+
         db.add(postModel)
         db.commit()
         db.refresh(postModel)
         return postModel
     
     def update_post(self, db:Session, 
-                    post_id: int, post: PostReq) -> tuple[Post|None,RESULT_CODE]:
+                    post_id: int, post: PostReq, userDict: dict) -> tuple[Post|None,RESULT_CODE]:
         oldPost = db.get(Post, post_id)
         if not oldPost:
             return (None,RESULT_CODE.NOT_FOUND)
         
-        # user_id = userDict.get('id', 0)
+        user_id = userDict.get('id', 0)
 
-        # if user_id != oldPost.author_id:
-        #     return (None, RESULT_CODE.FAILED)
+        if user_id != oldPost.author_id:
+            return (None, RESULT_CODE.USER_FAIL)
         
         dictToUpdate = asdict(post)
         oldPost.sqlmodel_update(dictToUpdate)
@@ -53,10 +57,13 @@ class PostService:
             return (None, RESULT_CODE.FAILED)
         return (oldPost, RESULT_CODE.SUCCESS)
     
-    def delete_post(self, db: Session, post_id: int) -> RESULT_CODE:
+    def delete_post(self, db: Session, post_id: int, userDict: dict) -> RESULT_CODE:
         post = db.get(Post, post_id)
         if not post:
-            return RESULT_CODE.NOT_FOUND 
+            return RESULT_CODE.NOT_FOUND
+        user_id = userDict.get('id', 0)
+        if post.author_id != user_id:
+            return RESULT_CODE.USER_FAIL
         try:
             db.delete(post)
             db.commit()
@@ -65,10 +72,15 @@ class PostService:
         return RESULT_CODE.SUCCESS
     
     def pullup_post(self, db:Session, 
-                        post_id: int) -> tuple[Post|None,RESULT_CODE]:
+                        post_id: int, userDict:dict) -> tuple[Post|None,RESULT_CODE]:
             oldPost = db.get(Post, post_id)
             if not oldPost:
                 return (None,RESULT_CODE.NOT_FOUND)
+            
+            post = db.get(Post, post_id)
+            user_id = userDict.get('id', 0)
+            if post.author_id != user_id:
+                return RESULT_CODE.USER_FAIL
 
             dictToUpdate = {
                 'updated_at': int(time.time())
